refactor(daemon): move project owners to runtime crates
