+++ crates/tracedecay-code-index-runtime/src/lib.rs
@@ -57,6 +57,7 @@ pub mod git_watch;
 pub(crate) mod logging;
 pub mod mcp_admission;
 pub(crate) mod ports;
+pub mod project_reads;
 pub mod semantic_activation_reconciler;
 pub(crate) mod semantic_code;
 pub mod semantic_evaluation;
