+++ crates/tracedecay-daemon-service/src/lib.rs
@@ -51,6 +51,7 @@
 pub use tracedecay_runtime_core::DAEMON_TASK_ABORT_DEADLINE as TASK_ABORT_DEADLINE;
 
 pub mod invocation;
+pub mod project_owner_registration;
 pub mod project_runtime;
 pub mod request_cancellation;
