+++ crates/tracedecay/src/daemon/project_open_owners.rs
@@ -8,45 +8,36 @@
 use std::collections::BTreeSet;
 use std::path::Path;
 use std::sync::Arc;
-use std::sync::atomic::{AtomicU8, Ordering};
 use std::time::{Duration, Instant};
 
 use tracedecay_application::advisory::GitHubRepositoryTargetV1;
-use tracedecay_contracts::feedback::{
-    CI_FAILURE_LOCALIZE_CAPABILITY_ID_V1, FEEDBACK_DIAGNOSTICS_CAPABILITY_ID_V1,
-    FEEDBACK_EXPAND_CAPABILITY_ID_V1, FEEDBACK_GET_CAPABILITY_ID_V1,
-    FEEDBACK_LIST_CAPABILITY_ID_V1, GITHUB_REVIEW_INGEST_CAPABILITY_ID_V1,
-    PROXIMITY_CAPABILITY_ID_V1,
+use tracedecay_application::project_open_authorization::{
+    ProjectOpenSourceAccessAuthorityV1, project_open_work_grant,
 };
 use tracedecay_contracts::{ApplicationContractError, ResolvedScope, now_micros};
-use tracedecay_domain::configuration::{
-    ACCESS_RULES_SETTING_KEY, AuthorityRef, CapabilityResolutionContextV1, ConfigurationValueV1,
-    SOURCE_BINDINGS_SETTING_KEY, ScopeSourceBinding, SettingKey, SourceBindingId, SourceKindV1,
-    resolve_restrictive_capabilities,
-};
 use tracedecay_domain::feedback::GitHubPullRequestIdV1;
-use tracedecay_domain::{
-    ActorId, CapabilityId as DomainCapabilityId, LocatorDigest, ProjectId, UtcMicros,
-    canonical_sha256,
-};
-use tracedecay_tool_catalog::{CapabilityId, UseCaseId};
+use tracedecay_domain::{ActorId, ProjectId, UtcMicros, canonical_sha256};
 
 use super::DaemonInvocationState;
-use tracedecay_contracts::request_identity::{PreviewIdentityDomain, derive_preview_identity};
-
-const SOURCE_EDIT_PRIVACY_KEY_EPOCH_V1: u64 = 1;
 use crate::daemon::callable_code_authorization::DaemonCallableCodeAuthorizationSource;
 use crate::mcp::McpServer;
 use tracedecay_application::lsp_runtime::DaemonLspSessionFactory;
-use tracedecay_application::primitives::{
-    admitted_root_uri_for_project, locator_digest_for_project,
+use tracedecay_application::primitives::admitted_root_uri_for_project;
+use tracedecay_application::semantic_runtime::{
+    InitialSemanticActivationRestoreV1, ProjectSemanticActivationExt,
+    classify_initial_semantic_activation_restore,
 };
-use tracedecay_application::semantic_runtime::ProjectSemanticActivationExt;
-use tracedecay_application::source_authorization::ProjectSourceAccessSnapshot;
+use tracedecay_application::source_authorization::{
+    ProjectSourceAccessSnapshot, ProjectSourceAccessSnapshotPort,
+};
 use tracedecay_code_index_runtime::git_transactions::DaemonGitIndexTransactionServiceRegistry;
 use tracedecay_daemon_service::{
     DaemonContextScoutRuntimeRegistrationError, DaemonFeedbackRuntimeRegistrationError,
     DaemonNativeIntegrationRuntimeRegistrar, DaemonWorkProposalRoutingAuthorityV1,
+    project_owner_registration::{
+        ProjectSourceEditAuthorizationV1, ProjectSourceEditOwnerV1, SourceEditMutationGate,
+        production_lsp_registration, project_open_lsp_scope_grant, project_owner_capabilities,
+    },
 };
 use tracedecay_domain::errors::{Result, TraceDecayError};
 use tracedecay_lsp::analyzer::broker::AdmittedLspProvider;
@@ -54,59 +45,21 @@ use tracedecay_lsp::analyzer::client::LspRefreshTimeouts;
 
 mod advisory_runtime;
 mod automation_effect_recovery;
-mod code_index_reads;
-mod lsp_registration;
+#[cfg(test)]
+#[path = "project_open_owners/code_index_reads/ignored_dependency_admission_tests.rs"]
+mod code_index_ignored_dependency_admission_tests;
 mod primitive_runtime;
 mod query_authority_upgrade;
-mod source_edit_owner;
-#[cfg(test)]
-mod work_grant_tests;
 
 pub(crate) use advisory_runtime::ProjectOpenDependentOwnerState;
 pub(super) use advisory_runtime::{
     register_project_open_dependent_owners, spawn_semantic_owner_registration,
 };
 pub(crate) use automation_effect_recovery::reconcile_project_open_automation_effects;
-pub(crate) use code_index_reads::{
-    project_code_graph_projection_read_port, project_code_index_generation_census_reader,
-    project_code_index_ignored_dependency_admission_port,
-};
 
-use lsp_registration::production_lsp_registration;
 use primitive_runtime::open_and_register_project_primitive_runtime;
-use source_edit_owner::{
-    install_project_open_source_edit_rollback_owner, source_edit_authority_error,
-    source_edit_contract_error, source_edit_request_context, source_edit_surface_result,
-};
-
-/// Whether this route's code index is disabled by contract, so no generation
-/// will ever be published for it.
-///
-/// `f347a0a46` gates project-open code-index activation for a linked worktree
-/// behind `sync.watch_linked_worktrees`, which defaults off. Such a route
-/// serves, but never indexes: a deferred owner that waits for its first
-/// generation waits for the daemon's whole life. That wait is not idle. Both
-/// deferred owners also wake on the *global* serving-seat signal, so every
-/// publication another route makes re-enters their mount attempt, and each
-/// attempt takes the project store writer lane this route shares with the
-/// admitted one — the lane a concurrently opening sibling and Context Scout
-/// durable startup are both waiting on. Answer that wait with the typed
-/// disabled state at spawn time instead of parking a task that can only ever
-/// contend.
-pub(super) fn code_index_disabled_for_scope(
-    invocation: &DaemonInvocationState,
-    scope: &ResolvedScope,
-) -> bool {
-    invocation
-        .code_index_schedulers
-        .automatic_admission_for_scope(scope)
-        == Some(
-            tracedecay_code_index_runtime::code_index_scheduler::CodeIndexAutomaticAdmissionV1::LinkedWorktreeDisabled,
-        )
-}
 
 const DAEMON_REQUESTER: &str = "actor.tracedecay-daemon.project-open";
-const DAEMON_BINDING: &str = "binding.tracedecay-daemon.project-open";
 const GRANT_HORIZON: Duration = Duration::from_hours(24);
 const POLICY_REVISION_V1: u64 = 1;
 const LSP_DIAGNOSTICS_QUIET: Duration = Duration::from_secs(2);
@@ -114,434 +67,69 @@ pub(super) use tracedecay_daemon_service::{
     LSP_WORKSPACE_CAPABILITY_ID_V1, LSP_WORKSPACE_USE_CASE_ID_V1,
 };
 
-#[derive(Clone)]
-struct ProjectOpenSourceEditAuthorizationV1 {
-    project_root: std::path::PathBuf,
-    scope: ResolvedScope,
-    configuration: Arc<tracedecay_configuration::ProjectConfigurationRuntime>,
-}
-
-struct CurrentSourceEditAuthorityV1 {
-    receipt: tracedecay_contracts::AuthorityReceipt,
-    proof: tracedecay_contracts::SourceEditEffectProofV1,
-}
-
-impl ProjectOpenSourceEditAuthorizationV1 {
-    #[hotpath::skip]
-    async fn current_access(
-        &self,
-        observed_at: UtcMicros,
-    ) -> std::result::Result<ProjectSourceAccessSnapshot, tracedecay_contracts::ApplicationProblem>
-    {
-        let current = self
-            .configuration
-            .client()
-            .current()
-            .await
-            .map_err(|_| concealed_source_edit_problem())?;
-        daemon_owned_project_source_access_at(
-            &self.scope,
-            &self.project_root,
-            &current,
-            observed_at,
-        )
-        .map_err(|_| concealed_source_edit_problem())
-    }
-
-    #[hotpath::skip]
-    async fn current_authority(
-        &self,
-        context: &tracedecay_contracts::RequestContext,
-        operation: &tracedecay_contracts::ApplicationOperation,
-        observed_at: UtcMicros,
-    ) -> std::result::Result<CurrentSourceEditAuthorityV1, tracedecay_contracts::ApplicationProblem>
-    {
-        let access = self.current_access(observed_at).await?;
-        if context.admission_at(observed_at) != tracedecay_contracts::RequestAdmission::Admitted
-            || !access.allows(context, operation, observed_at)
-        {
-            return Err(concealed_source_edit_problem());
-        }
-        let catalog = crate::catalog_composition::build_application_catalog_snapshot()
-            .map_err(|_| concealed_source_edit_problem())?;
-        let manifest = catalog
-            .capability(operation.capability_id())
-            .ok_or_else(concealed_source_edit_problem)?;
-        let catalog_digest = tracedecay_domain::ManifestDigest::new(catalog.digest().to_string())
-            .map_err(|_| concealed_source_edit_problem())?;
-        let privacy_domain_id = tracedecay_domain::PrivacyDomainId::new(format!(
-            "privacy.local-source-edit.{}",
-            access.scope.project_id.as_str()
-        ))
-        .map_err(|_| concealed_source_edit_problem())?;
-        let privacy_digest = canonical_sha256(&(
-            "tracedecay.daemon.source-edit-privacy.v1",
-            &privacy_domain_id,
-            SOURCE_EDIT_PRIVACY_KEY_EPOCH_V1,
-            manifest.privacy(),
-            manifest.denied_disclosure(),
-            manifest.scope(),
-            &access.binding,
-            &access.configuration_provenance_digest,
-        ))
-        .map_err(|_| concealed_source_edit_problem())?;
-        let policy_digest = canonical_sha256(&(
-            "tracedecay.daemon.source-edit-policy.v1",
-            &access.scope,
-            &access.requester,
-            &access.binding,
-            &access.configuration_digest,
-            &access.configuration_provenance_digest,
-            operation.capability_id(),
-            operation.use_case_id(),
-            &catalog_digest,
-            &privacy_digest,
-        ))
-        .map_err(|_| concealed_source_edit_problem())?;
-        let policy = tracedecay_contracts::PolicyDecisionRef::new(
-            "policy.daemon.source-edit.v1",
-            POLICY_REVISION_V1,
-            policy_digest,
-            tracedecay_domain::ComponentVersion::new("tracedecay.daemon.source-edit-policy.v1")
-                .map_err(|_| concealed_source_edit_problem())?,
-        )
-        .map_err(|_| concealed_source_edit_problem())?;
-        let receipt =
-            tracedecay_contracts::AuthorityReceipt::from_context(context, policy, observed_at)
-                .map_err(|_| concealed_source_edit_problem())?;
-        let proof = tracedecay_contracts::SourceEditEffectProofV1 {
-            policy_digest: receipt.policy.digest.clone(),
-            configuration_revision_id: access.configuration_revision,
-            configuration_digest: access.configuration_digest,
-            catalog_revision: manifest.routing().revision(),
-            catalog_digest,
-            privacy_domain_id,
-            privacy_key_epoch: SOURCE_EDIT_PRIVACY_KEY_EPOCH_V1,
-            privacy_digest,
-            external_proof: None,
-        };
-        proof
-            .validate_for(&receipt)
-            .map_err(|_| concealed_source_edit_problem())?;
-        Ok(CurrentSourceEditAuthorityV1 { receipt, proof })
-    }
-}
-
-impl tracedecay_contracts::SourceEditAuthorizationPort for ProjectOpenSourceEditAuthorizationV1 {
-    fn admit<'a>(
-        &'a self,
-        context: &'a tracedecay_contracts::RequestContext,
-        operation: &'a tracedecay_contracts::ApplicationOperation,
-        observed_at: UtcMicros,
-    ) -> tracedecay_contracts::SourceEditAuthorizationFuture<'a> {
-        Box::pin(async move {
-            self.current_authority(context, operation, observed_at)
-                .await
-                .and_then(|current| {
-                    tracedecay_contracts::SourceEditAuthorizationAdmissionV1::new(
-                        current.receipt,
-                        current.proof,
-                        context.scope(),
-                    )
-                    .map_err(|_| concealed_source_edit_problem())
-                })
-        })
-    }
-
-    fn recheck_effect<'a>(
-        &'a self,
-        context: &'a tracedecay_contracts::RequestContext,
-        operation: &'a tracedecay_contracts::ApplicationOperation,
-        admission: &'a tracedecay_contracts::SourceEditAuthorizationAdmissionV1,
-        observed_at: UtcMicros,
-    ) -> tracedecay_contracts::SourceEditAuthorizationFuture<'a> {
-        Box::pin(async move {
-            let current = self
-                .current_authority(context, operation, observed_at)
-                .await?;
-            if current.receipt.grant_id != admission.receipt.grant_id
-                || current.receipt.grant_revision != admission.receipt.grant_revision
-                || current.receipt.grant_digest != admission.receipt.grant_digest
-                || current.receipt.authorized_scope_digest
-                    != admission.receipt.authorized_scope_digest
-                || current.receipt.disclosure != admission.receipt.disclosure
-                || current.receipt.policy != admission.receipt.policy
-                || current.proof != admission.proof
-            {
-                return Err(concealed_source_edit_problem());
-            }
-            tracedecay_contracts::SourceEditAuthorizationAdmissionV1::new(
-                current.receipt,
-                current.proof,
-                context.scope(),
-            )
-            .map_err(|_| concealed_source_edit_problem())
-        })
-    }
-}
-
-fn concealed_source_edit_problem() -> tracedecay_contracts::ApplicationProblem {
-    tracedecay_contracts::ApplicationProblem::not_found_or_not_authorized(
-        tracedecay_contracts::RetryDirective::Never,
-    )
-}
-
-async fn invoke_project_open_source_edit(
-    graph: Arc<crate::tracedecay::TraceDecay>,
-    code_graph: Arc<dyn tracedecay_graph_query::CodeGraphProjectionReadPort>,
-    authorization: ProjectOpenSourceEditAuthorizationV1,
-    invocation: crate::mcp::server::SourceEditInvocationV1,
-) -> Result<tracedecay_contracts::source_edit::SourceEditSurfaceResultV1> {
-    let observed_at = now_micros();
-    let operation = tracedecay_contracts::source_edit_operation(invocation.edit.kind())
-        .map_err(source_edit_contract_error)?;
-    let access = authorization
-        .current_access(observed_at)
-        .await
-        .map_err(|_| source_edit_authority_error())?;
-    let context = source_edit_request_context(
-        &access,
-        invocation.request_id,
-        &operation,
-        observed_at,
-        invocation.deadline.clone(),
-        invocation.cancellation.context(),
-    )?;
-    let effect_control = tracedecay_source_edit::SourceEditEffectControlV1::new(
-        context.deadline().clone(),
-        invocation.cancellation.clone(),
-    );
-    let current = authorization
-        .current_authority(&context, &operation, observed_at)
-        .await
-        .map_err(|_| source_edit_authority_error())?;
-    let dry_run = invocation.edit.dry_run();
-    let idempotency_key = match invocation.idempotency_key {
-        Some(key) => key,
-        None if dry_run => {
-            let preview_identity = derive_preview_identity(
-                PreviewIdentityDomain::SourceEdit,
-                context.request_id(),
-                &invocation.edit,
-            )
-            .map_err(|error| TraceDecayError::Config {
-                message: format!("source edit preview identity failed: {error}"),
-            })?;
-            tracedecay_contracts::IdempotencyKey::new(format!("preview.{preview_identity}"))
-                .map_err(source_edit_contract_error)?
-        }
-        None => {
-            return Err(TraceDecayError::Config {
-                message: "source edit apply requires an idempotency key".to_owned(),
-            });
-        }
-    };
-    let expected_state = match invocation.expected_state {
-        Some(state) => state,
-        None if dry_run => canonical_sha256(&(
-            "tracedecay.source-edit-preview-unbound-state.v1",
-            context.request_id(),
-            &invocation.edit,
-        ))
-        .map_err(|error| TraceDecayError::Config {
-            message: format!("source edit preview state identity failed: {error}"),
-        })?,
-        None => {
-            return Err(TraceDecayError::Config {
-                message: "source edit apply requires an expected state".to_owned(),
-            });
-        }
-    };
-    let request = tracedecay_contracts::SourceEditEffectRequestV1 {
-        context,
-        authority: current.receipt.clone(),
-        edit: invocation.edit,
-        idempotency_key,
-        expected_state,
-        proof: current.proof,
-        observed_at,
-    };
-    tracedecay_source_edit::execute_source_edit_with_control(
-        &*graph,
-        code_graph.as_ref(),
-        &operation,
-        request,
-        &authorization,
-        &effect_control,
-    )
-    .await
-    .and_then(source_edit_surface_result)
-}
-
-async fn invoke_project_open_source_edit_reconciliation(
-    graph: Arc<crate::tracedecay::TraceDecay>,
-    authorization: ProjectOpenSourceEditAuthorizationV1,
-    invocation: crate::mcp::server::SourceEditReconciliationInvocationV1,
-) -> Result<tracedecay_contracts::source_edit::SourceEditSurfaceResultV1> {
-    let observed_at = now_micros();
-    let effect_control = tracedecay_source_edit::SourceEditEffectControlV1::new(
-        invocation.deadline.clone(),
-        invocation.cancellation.clone(),
-    );
-    let operation = tracedecay_contracts::source_edit_reconciliation_operation()
-        .map_err(source_edit_contract_error)?;
-    let access = authorization
-        .current_access(observed_at)
-        .await
-        .map_err(|_| source_edit_authority_error())?;
-    let context = source_edit_request_context(
-        &access,
-        invocation.request_id,
-        &operation,
-        observed_at,
-        invocation.deadline,
-        invocation.cancellation.context(),
-    )?;
-    let current = authorization
-        .current_authority(&context, &operation, observed_at)
-        .await
-        .map_err(|_| source_edit_authority_error())?;
-    let request = tracedecay_contracts::SourceEditReconciliationRequestV1 {
-        context,
-        authority: current.receipt.clone(),
-        kind: invocation.kind,
-        effect_id: invocation.effect_id,
-        idempotency_key: invocation.idempotency_key,
-        attempt_idempotency_key: invocation.attempt_idempotency_key,
-        input_digest: invocation.input_digest,
-        disposition: invocation.disposition,
-        proof: current.proof,
-        observed_at,
-    };
-    tracedecay_source_edit::reconcile_source_edit_effect_unknown_with_control(
-        &*graph,
-        request,
-        &authorization,
-        &effect_control,
-    )
-    .await
-    .and_then(source_edit_surface_result)
-}
-
-/// Publication state of the daemon-owned source-edit mutation lane.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum SourceEditMutationState {
-    /// Owner registration is still mounting the exact mutation authority. A
-    /// caller that retries later can succeed.
-    Warming,
-    /// The mutation authority is published.
-    Ready,
-    /// Owner registration failed. This server's mutation lane never opens, so
-    /// retrying the same request against it cannot succeed.
-    Failed,
-}
-
-/// Gates source-edit mutations on the state of their daemon-owned authority.
-///
-/// The preview executors are installed with the read-only core, so a mutation
-/// request can arrive before owner registration reaches the Git transaction
-/// authority — or after registration failed outright. Both cases fail closed,
-/// but only the first is retryable: reporting a failed publication as "warming"
-/// invites a caller to retry a lane that will never open.
-#[derive(Debug)]
-pub(crate) struct SourceEditMutationGate {
-    state: AtomicU8,
-}
-
-impl SourceEditMutationGate {
-    const WARMING: u8 = 0;
-    const READY: u8 = 1;
-    const FAILED: u8 = 2;
-
-    pub(crate) fn warming() -> Arc<Self> {
-        Arc::new(Self {
-            state: AtomicU8::new(Self::WARMING),
-        })
-    }
-
-    #[cfg(feature = "test-transport")]
-    pub(crate) fn ready() -> Arc<Self> {
-        Arc::new(Self {
-            state: AtomicU8::new(Self::READY),
-        })
-    }
-
-    pub(crate) fn state(&self) -> SourceEditMutationState {
-        match self.state.load(Ordering::Acquire) {
-            Self::READY => SourceEditMutationState::Ready,
-            Self::FAILED => SourceEditMutationState::Failed,
-            _ => SourceEditMutationState::Warming,
-        }
-    }
-
-    pub(crate) fn mark_ready(&self) {
-        self.state.store(Self::READY, Ordering::Release);
-    }
-
-    /// Retires the mutation lane. A publication that failed after opening the
-    /// lane still retires the whole server, so this overwrites `Ready` rather
-    /// than leaving mutations authorized against a server being torn down.
-    pub(crate) fn mark_failed(&self) {
-        self.state.store(Self::FAILED, Ordering::Release);
-    }
-
-    pub(crate) fn authorize_mutation(&self, lane: &str) -> Result<()> {
-        match self.state() {
-            SourceEditMutationState::Ready => Ok(()),
-            SourceEditMutationState::Warming => Err(TraceDecayError::Config {
-                message: format!("daemon-owned source edit {lane} authority is warming"),
-            }),
-            SourceEditMutationState::Failed => Err(TraceDecayError::Config {
-                message: format!(
-                    "daemon-owned source edit {lane} authority failed to publish; reopen the project"
-                ),
-            }),
-        }
-    }
-}
-
 fn install_project_open_source_edit_owners(
     server: &McpServer,
-    graph: Arc<crate::tracedecay::TraceDecay>,
-    code_graph: Arc<dyn tracedecay_graph_query::CodeGraphProjectionReadPort>,
-    authorization: ProjectOpenSourceEditAuthorizationV1,
-    mutation: Arc<SourceEditMutationGate>,
+    owner: Arc<ProjectSourceEditOwnerV1>,
 ) -> Result<()> {
-    let source_edit_graph = Arc::clone(&graph);
-    let source_edit_code_graph = Arc::clone(&code_graph);
-    let source_edit_reconciliation_authorization = authorization.clone();
-    let source_edit_rollback_authorization = authorization.clone();
-    let source_edit_mutation = Arc::clone(&mutation);
+    let edit_owner = Arc::clone(&owner);
     server
         .install_source_edit_executor(Arc::new(move |request| {
-            let graph = Arc::clone(&source_edit_graph);
-            let code_graph = Arc::clone(&source_edit_code_graph);
-            let authorization = authorization.clone();
-            let mutation = Arc::clone(&source_edit_mutation);
+            let owner = Arc::clone(&edit_owner);
             Box::pin(async move {
-                if !request.edit.dry_run() {
-                    mutation.authorize_mutation("mutation")?;
-                }
-                invoke_project_open_source_edit(graph, code_graph, authorization, request).await
+                owner
+                    .execute(
+                        request.edit,
+                        request.idempotency_key,
+                        request.expected_state,
+                        request.request_id,
+                        request.deadline,
+                        request.cancellation,
+                    )
+                    .await
             })
         }))
         .map_err(|_| TraceDecayError::Config {
             message: "project-open source edit authority was already installed".to_owned(),
         })?;
-    install_project_open_source_edit_rollback_owner(
-        server,
-        Arc::clone(&graph),
-        source_edit_rollback_authorization,
-        Arc::clone(&mutation),
-    )?;
+    let rollback_owner = Arc::clone(&owner);
+    server
+        .install_source_edit_rollback_executor(Arc::new(move |request| {
+            let owner = Arc::clone(&rollback_owner);
+            Box::pin(async move {
+                owner
+                    .rollback(
+                        request.effect_id,
+                        request.original_idempotency_key,
+                        request.idempotency_key,
+                        request.original_input_digest,
+                        request.expected_state,
+                        request.request_id,
+                        request.deadline,
+                        request.cancellation,
+                    )
+                    .await
+            })
+        }))
+        .map_err(|_| TraceDecayError::Config {
+            message: "project-open source edit rollback authority was already installed".to_owned(),
+        })?;
     server
         .install_source_edit_reconciliation_executor(Arc::new(move |request| {
-            let graph = Arc::clone(&graph);
-            let authorization = source_edit_reconciliation_authorization.clone();
-            let mutation = Arc::clone(&mutation);
+            let owner = Arc::clone(&owner);
             Box::pin(async move {
-                mutation.authorize_mutation("reconciliation")?;
-                invoke_project_open_source_edit_reconciliation(graph, authorization, request).await
+                owner
+                    .reconcile(
+                        request.kind,
+                        request.effect_id,
+                        request.idempotency_key,
+                        request.attempt_idempotency_key,
+                        request.input_digest,
+                        request.disposition,
+                        request.request_id,
+                        request.deadline,
+                        request.cancellation,
+                    )
+                    .await
             })
         }))
         .map_err(|_| TraceDecayError::Config {
@@ -568,19 +156,32 @@ pub(crate) async fn install_project_open_source_edit_preview_owner(
             .map_err(|error| TraceDecayError::Config {
                 message: format!("project-open source edit preview scope denied: {error}"),
             })?;
-    let authorization = ProjectOpenSourceEditAuthorizationV1 {
-        project_root: project_root.to_path_buf(),
+    let catalog =
+        crate::catalog_composition::build_application_catalog_snapshot().map_err(|error| {
+            TraceDecayError::Config {
+                message: format!("project-open source edit catalog is unavailable: {error}"),
+            }
+        })?;
+    let authorization = ProjectSourceEditAuthorizationV1::new(
+        project_root.to_path_buf(),
         scope,
-        configuration: Arc::clone(graph.configuration_runtime()),
-    };
+        Arc::clone(graph.configuration_runtime()),
+        Arc::new(catalog),
+        Arc::new(project_open_source_access_authority().map_err(|error| {
+            TraceDecayError::Config {
+                message: format!("project-open source edit access authority is invalid: {error}"),
+            }
+        })?),
+    );
+    let runtime: Arc<dyn tracedecay_source_edit::SourceEditRuntimePort> = graph;
     let mutation = SourceEditMutationGate::warming();
-    install_project_open_source_edit_owners(
-        server,
-        graph,
+    let owner = Arc::new(ProjectSourceEditOwnerV1::new(
+        runtime,
         code_graph,
         authorization,
         Arc::clone(&mutation),
-    )?;
+    ));
+    install_project_open_source_edit_owners(server, owner)?;
     Ok(mutation)
 }
 
@@ -604,23 +205,15 @@ pub(crate) async fn install_project_open_source_edit_owners_for_test(
         .configuration_target()
         .project_id
         .clone();
-    let scope =
-        tracedecay_code_index_runtime::resolved_scope_for_project(&project_root, &project_id)
-            .map_err(|error| TraceDecayError::Config {
-                message: format!("test project-open resolved scope denied: {error}"),
-            })?;
-    let authorization = ProjectOpenSourceEditAuthorizationV1 {
-        project_root,
-        scope,
-        configuration: Arc::clone(graph.configuration_runtime()),
-    };
-    install_project_open_source_edit_owners(
+    let mutation = install_project_open_source_edit_preview_owner(
         server,
         graph,
         code_graph,
-        authorization,
-        SourceEditMutationGate::ready(),
-    )?;
+        &project_root,
+        project_id.as_str(),
+    )
+    .await?;
+    mutation.mark_ready();
     Ok(true)
 }
 
@@ -1208,35 +801,6 @@ pub(super) async fn register_project_open_production_owners(
     })
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum InitialSemanticActivationRestoreV1 {
-    Mounted,
-    Deferred,
-}
-
-fn classify_initial_semantic_activation_restore(
-    observed: std::result::Result<
-        (),
-        tracedecay_application::semantic_runtime::RetrievalProfileActivationObserverErrorV1,
-    >,
-) -> std::result::Result<
-    InitialSemanticActivationRestoreV1,
-    tracedecay_application::semantic_runtime::RetrievalProfileActivationObserverErrorV1,
-> {
-    use tracedecay_application::semantic_runtime::RetrievalProfileActivationObserverErrorV1;
-
-    match observed {
-        Ok(()) => Ok(InitialSemanticActivationRestoreV1::Mounted),
-        Err(RetrievalProfileActivationObserverErrorV1::Unavailable) => {
-            Ok(InitialSemanticActivationRestoreV1::Deferred)
-        }
-        Err(
-            error @ (RetrievalProfileActivationObserverErrorV1::Rejected
-            | RetrievalProfileActivationObserverErrorV1::Conflict),
-        ) => Err(error),
-    }
-}
-
 #[hotpath::measure(label = "daemon.project.activate.semantic", future = true)]
 async fn register_semantic_configuration_owners(
     invocation: &DaemonInvocationState,
@@ -1718,209 +1282,26 @@ pub(super) fn daemon_owned_project_source_access_at(
     configuration: &tracedecay_configuration::config::PinnedRuntimeConfiguration,
     observed_at: UtcMicros,
 ) -> std::result::Result<ProjectSourceAccessSnapshot, ApplicationContractError> {
-    let locator = locator_digest_for_project(project_root)?;
-    let locator = LocatorDigest::new(locator.as_str().to_owned()).map_err(|_| {
-        ApplicationContractError::Inconsistent {
-            field: "project-open locator digest",
-        }
-    })?;
-    let binding = ScopeSourceBinding::new(
-        SourceBindingId::new(DAEMON_BINDING.to_owned()).map_err(|_| {
-            ApplicationContractError::Inconsistent {
-                field: "project-open source binding id",
-            }
-        })?,
-        SourceKindV1::Cursor,
-        locator,
-        AuthorityRef::Project(scope.project_id.clone()),
+    project_open_source_access_authority()?.source_access_at(
+        scope,
+        project_root,
+        configuration,
+        observed_at,
     )
-    .map_err(|_| ApplicationContractError::Inconsistent {
-        field: "project-open source binding",
-    })?;
-    if configuration.target().project_id != scope.project_id {
-        return Err(ApplicationContractError::Inconsistent {
-            field: "project-open configuration project",
-        });
-    }
-    configuration
-        .snapshot()
-        .validate()
-        .map_err(|_| ApplicationContractError::Inconsistent {
-            field: "project-open configuration snapshot",
-        })?;
+}
+
+fn project_open_source_access_authority()
+-> std::result::Result<ProjectOpenSourceAccessAuthorityV1, ApplicationContractError> {
     let requester = ActorId::new(DAEMON_REQUESTER.to_owned()).map_err(|_| {
         ApplicationContractError::Inconsistent {
             field: "project-open requester",
         }
     })?;
-    let authority = AuthorityRef::Project(scope.project_id.clone());
-    let bindings_key = SettingKey::new(SOURCE_BINDINGS_SETTING_KEY).map_err(|_| {
-        ApplicationContractError::Inconsistent {
-            field: "project-open source bindings key",
-        }
-    })?;
-    let Some(ConfigurationValueV1::SourceBindings(bindings)) =
-        configuration.snapshot().effective_values.get(&bindings_key)
-    else {
-        return Err(ApplicationContractError::Inconsistent {
-            field: "project-open source bindings",
-        });
-    };
-    let configured_bindings = bindings
-        .iter()
-        .filter(|candidate| {
-            candidate.source_kind == binding.source_kind && candidate.authority == authority
-        })
-        .collect::<Vec<_>>();
-    if configured_bindings.len() != 1
-        || configured_bindings.first().is_none_or(|candidate| {
-            candidate.source_locator_digest != binding.source_locator_digest
-        })
-    {
-        return Err(ApplicationContractError::Inconsistent {
-            field: "project-open source binding authority",
-        });
-    }
-    let binding = configured_bindings
-        .first()
-        .map(|configured| (**configured).clone())
-        .ok_or(ApplicationContractError::Inconsistent {
-            field: "project-open source binding authority",
-        })?;
-    let access_rules_key = SettingKey::new(ACCESS_RULES_SETTING_KEY).map_err(|_| {
-        ApplicationContractError::Inconsistent {
-            field: "project-open access rules key",
-        }
-    })?;
-    let Some(ConfigurationValueV1::AccessRules(access_rules)) = configuration
-        .snapshot()
-        .effective_values
-        .get(&access_rules_key)
-    else {
-        return Err(ApplicationContractError::Inconsistent {
-            field: "project-open access rules",
-        });
-    };
-    let granted_capabilities = production_owner_capabilities()?
-        .into_iter()
-        .map(|capability| DomainCapabilityId::new(capability.as_str().to_owned()))
-        .collect::<std::result::Result<BTreeSet<_>, _>>()
-        .map_err(|_| ApplicationContractError::Inconsistent {
-            field: "project-open granted capabilities",
-        })?;
-    let resolution = resolve_restrictive_capabilities(
-        granted_capabilities,
-        access_rules,
-        &CapabilityResolutionContextV1 {
-            actor: requester.clone(),
-            operation: None,
-            source_kind: binding.source_kind,
-            authority,
-            evaluated_at: observed_at,
-        },
-    )
-    .map_err(|_| ApplicationContractError::Inconsistent {
-        field: "project-open capability resolution",
-    })?;
-    let effective_capabilities = resolution
-        .effective
-        .into_iter()
-        .map(|capability| CapabilityId::new(capability.as_str().to_owned()))
-        .collect::<std::result::Result<BTreeSet<_>, _>>()
-        .map_err(|_| ApplicationContractError::Inconsistent {
-            field: "project-open effective capabilities",
-        })?;
-    Ok(ProjectSourceAccessSnapshot {
-        scope: scope.clone(),
+    Ok(ProjectOpenSourceAccessAuthorityV1::new(
         requester,
-        binding,
-        configuration_revision: configuration.revision_id().clone(),
-        configuration_digest: configuration.snapshot().effective_behavior_digest.clone(),
-        configuration_provenance_digest: configuration
-            .snapshot()
-            .resolution_provenance_digest
-            .clone(),
-        effective_capabilities,
-        grant_expires_at: UtcMicros(
-            observed_at
-                .0
-                .saturating_add(i64::try_from(GRANT_HORIZON.as_micros()).unwrap_or(i64::MAX)),
-        ),
-    })
-}
-
-pub(crate) struct DaemonOwnedProjectSourceAccess;
-
-impl tracedecay_application::ProjectSourceAccessSnapshotPort for DaemonOwnedProjectSourceAccess {
-    fn source_access_at(
-        &self,
-        scope: &ResolvedScope,
-        project_root: &Path,
-        configuration: &tracedecay_configuration::config::PinnedRuntimeConfiguration,
-        observed_at: UtcMicros,
-    ) -> std::result::Result<ProjectSourceAccessSnapshot, ApplicationContractError> {
-        daemon_owned_project_source_access_at(scope, project_root, configuration, observed_at)
-    }
-}
-
-fn project_open_work_grant(
-    access: &ProjectSourceAccessSnapshot,
-    observed_at: UtcMicros,
-) -> std::result::Result<tracedecay_contracts::CapabilityGrantSnapshot, ApplicationContractError> {
-    let capabilities = tracedecay_contracts::WORK_APPLICATION_OPERATION_IDS_V1
-        .iter()
-        .chain(tracedecay_contracts::WORKFLOW_APPLICATION_OPERATION_IDS.iter())
-        .chain(tracedecay_contracts::HANDOFF_APPLICATION_OPERATION_IDS_V1.iter())
-        .map(|(_, capability, _)| CapabilityId::new(*capability))
-        .collect::<std::result::Result<BTreeSet<_>, _>>()
-        .map_err(|_| ApplicationContractError::Inconsistent {
-            field: "project-open Work capabilities",
-        })?;
-    if observed_at >= access.grant_expires_at
-        || !capabilities
-            .iter()
-            .all(|capability| access.effective_capabilities.contains(capability))
-    {
-        return Err(ApplicationContractError::Inconsistent {
-            field: "project-open Work capability grant",
-        });
-    }
-    let use_cases = tracedecay_contracts::WORK_APPLICATION_OPERATION_IDS_V1
-        .iter()
-        .chain(tracedecay_contracts::WORKFLOW_APPLICATION_OPERATION_IDS.iter())
-        .chain(tracedecay_contracts::HANDOFF_APPLICATION_OPERATION_IDS_V1.iter())
-        .map(|(_, _, use_case)| tracedecay_tool_catalog::UseCaseId::new(*use_case))
-        .collect::<std::result::Result<BTreeSet<_>, _>>()
-        .map_err(|_| ApplicationContractError::Inconsistent {
-            field: "project-open Work use cases",
-        })?;
-    let grant_digest = canonical_sha256(&(
-        "tracedecay.project-open.work-grant.v1",
-        &access.scope,
-        &access.requester,
-        &access.binding,
-        &access.effective_capabilities,
-        &capabilities,
-        &use_cases,
+        project_owner_capabilities()?,
+        GRANT_HORIZON,
     ))
-    .map_err(|_| ApplicationContractError::Inconsistent {
-        field: "project-open Work grant digest",
-    })?;
-    tracedecay_contracts::CapabilityGrantSnapshot::new(
-        tracedecay_contracts::CapabilityGrantId::new(format!(
-            "grant.tracedecay-daemon.project-open.work.{}",
-            grant_digest.as_str().trim_start_matches("sha256:")
-        ))?,
-        POLICY_REVISION_V1,
-        grant_digest,
-        access.requester.clone(),
-        observed_at,
-        access.grant_expires_at,
-        access.scope.clone(),
-        capabilities,
-        use_cases,
-        tracedecay_contracts::DisclosureClass::Sensitive,
-    )
 }
 
 pub(super) fn project_open_retained_grant(
@@ -1977,277 +1358,6 @@ pub(super) fn project_open_retained_grant(
     )
 }
 
-pub(super) fn project_open_lsp_scope_grant(
-    access: &ProjectSourceAccessSnapshot,
-    observed_at: UtcMicros,
-) -> std::result::Result<tracedecay_contracts::CapabilityGrantSnapshot, ApplicationContractError> {
-    let capability = CapabilityId::new(LSP_WORKSPACE_CAPABILITY_ID_V1).map_err(|_| {
-        ApplicationContractError::Inconsistent {
-            field: "project-open LSP workspace capability",
-        }
-    })?;
-    if observed_at >= access.grant_expires_at
-        || !access.effective_capabilities.contains(&capability)
-    {
-        return Err(ApplicationContractError::Inconsistent {
-            field: "project-open LSP workspace capability grant",
-        });
-    }
-    let use_case = UseCaseId::new(LSP_WORKSPACE_USE_CASE_ID_V1).map_err(|_| {
-        ApplicationContractError::Inconsistent {
-            field: "project-open LSP workspace use case",
-        }
-    })?;
-    let capabilities = BTreeSet::from([capability]);
-    let use_cases = BTreeSet::from([use_case]);
-    let grant_digest = canonical_sha256(&(
-        "tracedecay.project-open.lsp-workspace-grant.v1",
-        &access.scope,
-        &access.requester,
-        &access.configuration_digest,
-        &access.configuration_provenance_digest,
-        &capabilities,
-        &use_cases,
-    ))
-    .map_err(|_| ApplicationContractError::Inconsistent {
-        field: "project-open LSP workspace grant digest",
-    })?;
-    tracedecay_contracts::CapabilityGrantSnapshot::new(
-        tracedecay_contracts::CapabilityGrantId::new(format!(
-            "grant.tracedecay-daemon.project-open.lsp-workspace.{}",
-            grant_digest.as_str().trim_start_matches("sha256:")
-        ))?,
-        POLICY_REVISION_V1,
-        grant_digest,
-        access.requester.clone(),
-        observed_at,
-        access.grant_expires_at,
-        access.scope.clone(),
-        capabilities,
-        use_cases,
-        tracedecay_contracts::DisclosureClass::Sensitive,
-    )
-}
-
-fn production_owner_capabilities()
--> std::result::Result<BTreeSet<CapabilityId>, ApplicationContractError> {
-    let mut capabilities = BTreeSet::new();
-    for capability in [
-        "capability.diagnostics.current",
-        FEEDBACK_DIAGNOSTICS_CAPABILITY_ID_V1,
-        FEEDBACK_GET_CAPABILITY_ID_V1,
-        FEEDBACK_EXPAND_CAPABILITY_ID_V1,
-        FEEDBACK_LIST_CAPABILITY_ID_V1,
-        "capability.application.feedback.impact",
-        "capability.application.feedback.affected-tests",
-        "capability.application.feedback.test-results",
-        "capability.application.code-query.exact-occurrence",
-        "capability.application.code-query.phrase-search",
-        "capability.application.code-query.callees",
-        "capability.application.code-query.facets",
-        "capability.application.code-query.timeline",
-        "capability.application.code-query.declaration",
-        "capability.application.code-query.definition",
-        "capability.application.code-query.type-definition",
-        "capability.application.code-query.references",
-        "capability.application.symbol-search",
-        GITHUB_REVIEW_INGEST_CAPABILITY_ID_V1,
-        CI_FAILURE_LOCALIZE_CAPABILITY_ID_V1,
-        PROXIMITY_CAPABILITY_ID_V1,
-        "capability.application.git.status",
-        "capability.application.git.diff",
-        "capability.application.git.history",
-        "capability.application.git.blame",
-        "capability.application.git.hunks",
-        LSP_WORKSPACE_CAPABILITY_ID_V1,
-        "capability.application.source-edit.ast-grep-rewrite",
-        "capability.application.source-edit.insert-at",
-        "capability.application.source-edit.insert-at-symbol",
-        "capability.application.source-edit.move-symbol",
-        "capability.application.source-edit.multi-str-replace",
-        "capability.application.source-edit.rename-symbol",
-        "capability.application.source-edit.replace-symbol",
-        "capability.application.source-edit.reconcile",
-        "capability.application.source-edit.rollback",
-        "capability.application.source-edit.str-replace",
-        "capability.git.stage-hunks",
-        "capability.git.unstage-hunks",
-        "capability.git.commit-index",
-    ] {
-        capabilities.insert(CapabilityId::new(capability.to_owned()).map_err(|_| {
-            ApplicationContractError::Inconsistent {
-                field: "project-open capability",
-            }
-        })?);
-    }
-    for descriptor in
-        tracedecay_contracts::retrieval::catalog::primitive_read_handler_descriptors()?
-    {
-        capabilities.insert(descriptor.operation().capability_id().clone());
-    }
-    for (_, capability, _) in tracedecay_contracts::WORK_APPLICATION_OPERATION_IDS_V1
-        .into_iter()
-        .chain(tracedecay_contracts::WORKFLOW_APPLICATION_OPERATION_IDS)
-        .chain(tracedecay_contracts::HANDOFF_APPLICATION_OPERATION_IDS_V1)
-    {
-        capabilities.insert(CapabilityId::new(capability).map_err(|_| {
-            ApplicationContractError::Inconsistent {
-                field: "project-open Work capability",
-            }
-        })?);
-    }
-    for operation in tracedecay_contracts::RetainedSurfaceOperation::CALLABLE {
-        let operation = tracedecay_contracts::retained_surface_application_operation(operation)?;
-        capabilities.insert(operation.capability_id().clone());
-    }
-    Ok(capabilities)
-}
-
-#[derive(Clone, Copy, Debug, Default)]
-pub(crate) struct DaemonCodeIndexScopeResolverV1;
-
-impl tracedecay_code_index_runtime::mcp_admission::CodeIndexScopeResolverV1
-    for DaemonCodeIndexScopeResolverV1
-{
-    fn resolved_scope_for_project(
-        &self,
-        project_root: &Path,
-        project_id: &ProjectId,
-    ) -> std::result::Result<
-        ResolvedScope,
-        tracedecay_code_index_runtime::mcp_admission::CodeIndexScopeUnavailableV1,
-    > {
-        tracedecay_code_index_runtime::resolved_scope_for_project(project_root, project_id)
-            .map_err(|_| tracedecay_code_index_runtime::mcp_admission::CodeIndexScopeUnavailableV1)
-    }
-}
-
 #[cfg(test)]
 #[path = "project_open_owners/git_catalog_tests.rs"]
 mod git_catalog_tests;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn transient_semantic_activation_restore_is_deferred_but_refusals_are_terminal() {
-        use tracedecay_application::semantic_runtime::RetrievalProfileActivationObserverErrorV1;
-
-        assert_eq!(
-            classify_initial_semantic_activation_restore(Err(
-                RetrievalProfileActivationObserverErrorV1::Unavailable,
-            )),
-            Ok(InitialSemanticActivationRestoreV1::Deferred),
-        );
-        for refusal in [
-            RetrievalProfileActivationObserverErrorV1::Rejected,
-            RetrievalProfileActivationObserverErrorV1::Conflict,
-        ] {
-            assert_eq!(
-                classify_initial_semantic_activation_restore(Err(refusal)),
-                Err(refusal),
-            );
-        }
-        assert_eq!(
-            classify_initial_semantic_activation_restore(Ok(())),
-            Ok(InitialSemanticActivationRestoreV1::Mounted),
-        );
-    }
-
-    #[test]
-    fn production_project_owner_grants_every_cataloged_git_read() {
-        let capabilities = production_owner_capabilities().expect("production capabilities");
-
-        for capability in [
-            "capability.application.git.status",
-            "capability.application.git.diff",
-            "capability.application.git.history",
-            "capability.application.git.blame",
-            "capability.application.git.hunks",
-        ] {
-            let capability = CapabilityId::new(capability).expect("Git read capability");
-            assert!(
-                capabilities.contains(&capability),
-                "{} must be granted to the daemon-owned project route",
-                capability.as_str()
-            );
-        }
-    }
-
-    #[test]
-    fn production_project_owner_grants_every_primitive_read() {
-        let capabilities = production_owner_capabilities().expect("production capabilities");
-
-        for descriptor in
-            tracedecay_contracts::retrieval::catalog::primitive_read_handler_descriptors()
-                .expect("primitive read descriptors")
-        {
-            assert!(
-                capabilities.contains(descriptor.operation().capability_id()),
-                "{} must be granted to the daemon-owned project route",
-                descriptor.operation().capability_id().as_str()
-            );
-        }
-    }
-
-    #[test]
-    fn production_project_owner_grants_every_retained_operation() {
-        let capabilities = production_owner_capabilities().expect("production capabilities");
-
-        for operation in tracedecay_contracts::RetainedSurfaceOperation::CALLABLE {
-            let operation = tracedecay_contracts::retained_surface_application_operation(operation)
-                .expect("retained application operation");
-            assert!(
-                capabilities.contains(operation.capability_id()),
-                "{} must be granted to the daemon-owned retained route",
-                operation.capability_id().as_str()
-            );
-        }
-    }
-
-    fn admitted(language: &str, analyzer_available: bool) -> AdmittedLspProvider {
-        AdmittedLspProvider {
-            language: language.to_owned(),
-            command: format!("{language}-language-server"),
-            analyzer_available,
-        }
-    }
-
-    #[test]
-    fn production_registration_mounts_dynamic_workspace_diagnostics_without_analyzer() {
-        let admitted = [admitted("rust", false)];
-        let (languages, gateway) = production_lsp_registration(&admitted);
-
-        assert_eq!(languages, vec!["rust"]);
-        assert!(gateway.supports_document_diagnostics);
-        assert!(gateway.supports_managed_diagnostics);
-        assert!(gateway.supports_workspace_diagnostics);
-        assert_eq!(
-            gateway.semantic,
-            tracedecay_lsp::SemanticCapability::ALL
-                .into_iter()
-                .collect()
-        );
-    }
-
-    #[test]
-    fn registration_preserves_every_admitted_project_language() {
-        for language in ["python", "typescript"] {
-            let admitted = [
-                admitted("rust", false),
-                admitted(language, true),
-                admitted("go", false),
-            ];
-            let (selected, gateway) = production_lsp_registration(&admitted);
-
-            assert_eq!(selected, vec!["rust", language, "go"]);
-            assert_eq!(
-                gateway.semantic,
-                tracedecay_lsp::SemanticCapability::ALL
-                    .into_iter()
-                    .collect()
-            );
-        }
-    }
-}
