+++ crates/tracedecay/src/daemon/project_open_owners/code_index_reads/ignored_dependency_admission_tests.rs
@@ -19,10 +19,10 @@ use tracedecay_domain::{
 };
 use tracedecay_tool_catalog::{CapabilityId, UseCaseId};
 
-use super::project_code_index_ignored_dependency_admission_port;
 use tracedecay_code_index_runtime::code_index_scheduler::{
     CodeGraphActivationPolicyV1, CodeIndexSchedulerRegistryV1, LatestCompleteCodeIndexV1,
 };
+use tracedecay_code_index_runtime::project_reads::project_code_index_ignored_dependency_admission_port;
 
 const PROJECT_ID: &str = "project.project-open-ignored-dependency";
