+++ crates/tracedecay/src/daemon/project_open_owners/primitive_runtime.rs
@@ -5,13 +5,11 @@ use std::sync::Arc;
 
 use tracedecay_application::primitives::{
     ProductionPrimitiveCodeAuthoritiesV1, ProductionPrimitiveOpenRequestV1,
-    open_production_primitive_runtime,
 };
 use tracedecay_application::source_authorization::ProjectSourceAccessSnapshot;
 
 use crate::daemon::DaemonInvocationState;
 use crate::mcp::McpServer;
-use tracedecay_daemon_service::DaemonPrimitiveRuntimeRegistrationError;
 use tracedecay_daemon_service::daemon_operation_event_authority;
 use tracedecay_domain::errors::{Result, TraceDecayError};
 use tracedecay_session_runtime::session_retrieval::DaemonSessionLookupPrimitiveV1;
@@ -41,47 +39,33 @@ pub(super) async fn open_and_register_project_primitive_runtime(
     let temporal = Arc::new(DaemonSessionLookupPrimitiveV1::new(
         server.project_session_application_retrieval_service(&access.scope)?,
     ));
-    let source = graph.source_read_context().ok_or_else(|| TraceDecayError::Config {
-        message: "project-open primitive runtime requires an exact registered source identity"
-            .to_owned(),
-    })?;
-    let primitive_runtime =
-        open_production_primitive_runtime(ProductionPrimitiveOpenRequestV1::new(
-            Arc::new(source),
-            ProductionPrimitiveCodeAuthoritiesV1 {
-                code_graph,
-                ignored_dependency_admission: Some(ignored_dependency_admission),
-                code_index: Arc::new(invocation.code_index_schedulers.clone()),
-                diagnostic_identity: Arc::new(invocation.code_index_schedulers.clone()),
-            },
-            session_db,
-            temporal,
-            access,
-            admitted_root_uri.to_owned(),
-            daemon_operation_event_authority(),
-        ))
+    let source = graph
+        .source_read_context()
+        .ok_or_else(|| TraceDecayError::Config {
+            message: "project-open primitive runtime requires an exact registered source identity"
+                .to_owned(),
+        })?;
+    invocation
+        .primitive_runtime_registrar()
+        .open_and_register(
+            project_root.to_path_buf(),
+            ProductionPrimitiveOpenRequestV1::new(
+                Arc::new(source),
+                ProductionPrimitiveCodeAuthoritiesV1 {
+                    code_graph,
+                    ignored_dependency_admission: Some(ignored_dependency_admission),
+                    code_index: Arc::new(invocation.code_index_schedulers.clone()),
+                    diagnostic_identity: Arc::new(invocation.code_index_schedulers.clone()),
+                },
+                session_db,
+                temporal,
+                access,
+                admitted_root_uri.to_owned(),
+                daemon_operation_event_authority(),
+            ),
+        )
         .await
         .map_err(|error| TraceDecayError::Config {
-            message: format!("project-open primitive runtime open failed: {error}"),
-        })?;
-    match invocation
-        .primitive_runtime_registrar()
-        .register(project_root.to_path_buf(), primitive_runtime)
-        .await
-    {
-        Ok(_) | Err(DaemonPrimitiveRuntimeRegistrationError::AlreadyRegistered) => Ok(()),
-        Err(DaemonPrimitiveRuntimeRegistrationError::RegistryClosed) => {
-            Err(TraceDecayError::Config {
-                message: "project-open primitive runtime registration failed: the daemon project runtime registry is closed"
-                    .to_owned(),
-            })
-        }
-        Err(DaemonPrimitiveRuntimeRegistrationError::ConcurrentBuildFailed { detail }) => {
-            Err(TraceDecayError::Config {
-                message: format!(
-                    "project-open primitive runtime registration failed after a concurrent build: {detail}"
-                ),
-            })
-        }
-    }
+            message: format!("project-open primitive runtime registration failed: {error}"),
+        })
 }
